Implement user preferences in VTS dashboard.
Implement user preferences for the main dashboard. Filter which
tests show up on dashboard home. Allow for users to subscribe to
email notifications individually via the web interface. This
functionality is added in a new preferences page that can handle
get and put requests.

Note, changes to the alert service will be made in a separate
CL.

Test: tested on local gcloud
Change-Id: I6e1a00abcd652a4e2ea3db5835b77cea6999593e