make local run using a config (without device info) work
this is when a test has a <test_name>.config file which does
not include the device info (e.g., serial number and test name).
note that that becomes common in vts host-side py test cases
because now the checked-in <test_name>.config is ready by the
VTS TF which adds device info and generates another
<test_name>.config file which is then consumed by the VTS py
runner. this change is for when a py test is invoked directly
from cmd-line without using the VTS TF. So the baseline_config
keeps the default device info and the auto-filled test name info
and is then merged to the <test_name>.config provided as the first
command-line argument.


Test: ./run-angler.sh
Change-Id: If11548da0c21cd50872db69b45eebda30f6cea01