Temporally disable recent LTP change.
A recent change in vts is incompatible with the old version of pinned vts test, which caused all vts presubmit to fail.
As the change involves removing vendor info which cannot be in the runner code anymore, that change cannot rollback.

Since it is a P0S0 bug, we need to push the newest build to production without completing the planned vts police cycle observation,
this change is to disable recent LTP change temporally to prevent potential presubmit test failure.

This change will be reverted.

Bug: 32477466
Test: local
Change-Id: I765b4c655a33c676053cd6ecb8b90213abb9d666