Add debugging params for viewing local VTS runs.
Add hidden flags so VTS devs can view local runs on the dashboard.
Flag disables filtering on device ID so that all results are
shown. No UI is present for this because non-VTS devs can't upload
to DB, so it would be confusing to have a UI element that has no
meaning to other users.

To use, just add &unfiltered= to URL when on the table view.

Test: local gcloud
Change-Id: I4274751329687fbd2894a5f9330b39fe3ccbd9d2