Auto generate default implementation for hidl callback interface
Hidl callback interface can contain hundreds of callback functions, while only a few are needed in a test case.

If the rest of the function implementation can be auto-generated with a dummy one, the test project development could be much easier.

This change make vts mirror auto generate a dummy implementation for each of the api that are not implemented. The dummy function will print out the api name and all the arguments received.

Bug: 34857249
Test: ENABLE_TREBLE=true make vts BUILD_GOOGLE_VTS=true -j32 && vts-tradefed run commandAndExit vts --skip-all-system-status-check --primary-abi-only --skip-preconditions --module VehicleHidlTest -l INFO
Change-Id: Ib2502562a83ba8b990050586ed93771ed61dfd26