Support declaring 64bit test source while running on 32bit device
Binary/gtest tests assume all the sources are available. However, there can be a case when a test be scheduled on both 32bit only and 32/64bit device.
In that case, if include both 32bit and 64bit binary sources in AndroidTest.xml, setupClass() will fail.
We should tolerate such a case and give test skip or warning.

This change removes the requirement that specified binary test source needs to exist on host.
If a source is specified for adb push but does not exist on host, a warning will be logged and the source entry will be ignored.


Bug: 34716596
Test: modified VibratorHidlTargetTest .xml file to include a gtest source that doesn't exist
Change-Id: I71981faee45d62bfcbdf0c0da54592589ccd4790