Fix determination of pre/post-submit test runs.
The device builds should determine whether the test is pre or
post-submit, except in the case when the test build is of type
OTHER (e.g. a locally build test run against a device). Otherwise,
presubmit runs could have a regular or presubmit test build (e.g. if
we pin the test build).

Test: local
Bug: 62038839
Change-Id: I6bca883c8e62b48a3d80674ab664e4ff24125cae