create HalHidlSetUp and HalHidlTearDown
This is a first setup, which created the virtual functions HalHidlSetUp and HalHidlTearDown.
SetUp and TearDown will remain virtual at this point.

After all HalHidl tests are converted to use HalHidlSetUp and HalHidlTearDown,
SetUp and TearDown will be set to final instead of virtual.

This allows the runner to perform extra tasks for tests and guaranteed not to be overridden.

Bug: 63105108
Test: make vts
Change-Id: Ib091688e22107319f58af2d0565c75a10b5d8130