Set default framework state to stopped.
Instead of having the framework on by default, set it to off when the
target is prepared. This does not change the API in the AndroidTest.xml
configuration, it just changes the default state. Tests which need the
framework off do not have to do anything. Tests which need the framework
on will turn it on at startup and off at teardown.

Test: run vts-codelab
Bug: 67783802
Change-Id: I7b4298c7a4b4cc276902f7c4a8d4a54d69af1ff7