Make gcov cleanup more robust.
Currently, gcov cleanup can fail if there is write contention (e.g. if a
HAL is flushing its coverage at the same time VTS is trying to cleanup).
This is unavoidable in some cases (e.g. if a hal frequently dlcloses a
lib) and should log a warning rather than stopping test execution.

Test: run vts-hal -m VtsHalLightV2_0Target
Bug: 67998360
Change-Id: Ia1dd0eda0af96edff2853b75f06503aab95f765b