Add support to allow custom config from AndroidTest.xml
Currently, if a developer wants to pass a user config option from AndroidTest.xml to the test module, it has to be defined in VtsMultideviceTest.java and base_test.py, which requires general vts framework side change for each parameters.

Therefore, they will have to use a separate config json file in project directory which is not recommended.

Developers should be able to define a config variable in AndroidTest.xml and receive in test module without having to change the framework.

Test: vts-tradefed run vts-selftest -m VtsSelfTestBaseTest -l INFO
Change-Id: Ib8555cd0d7ba663017e9ff197075ebccce8a9530
Fix: 71363536