Migrate host controller to test/framework
conceptually, host controller is not part of VTS
but part of VTS lab infrastructure. Since this tool is not
part of an infrastructure in a tranditional sense (but a
host-side tool), moving this to test/framework.

Test: mma
Bug: 63801212
Change-Id: I8629d64c708cd652aa760173fa8314d5ff934c53
(cherry picked from commit 27ac9fff6c0085337b1a0af4d3c64dbb4ff35954)