Set records' class name to module name in base class
In base_test, the test record's class name is set to Python class name,
while the filter's module name is set to the test-module-name in the
config file. The names mismatch when the user retries tests.

This patch replaces the class name with module name, which also changes
the test case names in the report and log.

Test: run vts -m VtsHalRadioV1_1Target
Bug: 71519627
Change-Id: I449bd26a4f6e0fd630eadb2626af93334edf3ef5
Merged-In: I449bd26a4f6e0fd630eadb2626af93334edf3ef5