Add vts-tradefed-standalone target and update vts-tradefed.
vts-tradefed-standalone will build vts-tradefed and the jars it
requires.

Updated vts-tradefed to look for the jars in the following manner:
- Check if tradefed jar is in VTS_TOOLS_OUT, then ANDROID_HOST_OUT/framework
- Look for the rest of the jars wherever it found the tradefed jar
- If it can't find tradefed jar in either path, just default to look for
  tradefed-prebuilt.jar and all the other jars in VTS_TOOLS_OUT.
This maintains existing vts-tradefed behavior when either built from vts
or in a zip while also enabling vts-tradefed to work when built using
vts-tradefed-standalone.

Bug: 73125127
Test: m vts-tradefed-standalone && vts-tradefed
m vts && vts-tradefed
Checked that vts-tradefed picked up jars in the expected JAR_DIR
depending on where it was built from (vts vs. vts-tradefed-standalone).

Change-Id: If59abdaec82dfb906bb530e0fd2b750c1efd54b4