Put VtsPythonVirtualenvPreparer in base plan level config
Since VtsPythonVirtualenvPreparer is used for every vts tests, it is better to put it in plan level config.
This reduces a great amount of test module preparation time when there are many modules in a plan.

If a module needs to install additional python dep, VtsPythonVirtualenvPreparer can be specified again in module level
config. Duplication is ok.

Test: run vts-selftest
Change-Id: I759df89ddc7d4771edfc9cab8931abba79272fc4