Fix vts-unit-tests and up vts unit tests quality
None of the failures looked related to change in
dependencies. Just bad and out of date unit tests.
- Ensure unit tests do not look for local path resources
when the resources will be inside the JAR.
- Clean leaked files
- Ensure mock verify is called to actually check the
callbacks.
- Fix some logic that was missing testEnded in a lot of
places probably generating some bad results.

Note: all unit tests pass with vts-unit-tests but not in
IDE, meaning some are assuming some environment which is
bad and should be fixed later.

Test: make vts, run vts-unit-tests
Bug: 76098093
Change-Id: I7c96b26822fd46fb6a06efe025d62df66bb384cc