Update vts_treble_vintf_test binary name.
Before 'vts_treble_vintf_test' meant 'ToT VtsTrebleVintf
test', but now it means 'legacy test for O-MR1'.

This change in meaning with the same name means that making
and running vts_treble_vintf_test no longer applies to modern
devices. Rather than running the legacy code in this case,
this will now cause an error so that the user can know that
the test has changed and then select the appropriate test.

Bug: 77983049
Bug: 77680266
Test: vts_treble_vintf_test_o_mr1
Change-Id: Ia26bd38785234d5e7e84b89172f2c69bf295d81a