Set done=false when setup_class fails
When setup_class fails, the runner adds an extra failure record. In
retry session, the record is neither removed nor retried.
This patch replaces the extra record with a field of error message.
If the error message is present, the result parser calls testRunFailed
to set done=false in test_result.xml. The whole module will be executed
again in the retry session.

Bug: 79436485
Bug: 78691622
Test: run vts -m VtsVndkDependency
Change-Id: I81d87e2ad40d28e668275b102b92f38dfdf44fe9