Move the netd integration test from vts-kernel to cts-on-gsi.
This test is not appropriate for VTS because it tests the
behaviour of netd, not the behaviour of kernel interfaces.
Thus, when upgrading a device, for example, from 8.1 to 9.0,
the 8.1 test's expectations are incorrect because the correct
behaviour changed when the OS was upgraded.

Bug: 77680167
Test: make -j64 vts vts-tradefed
Merged-In: Ia51874240d0be290bcb43a4f5a8f9cfa9e8d2112
Change-Id: Ia51874240d0be290bcb43a4f5a8f9cfa9e8d2112
(cherry picked from commit 4ff67aa1490090f48301a66911acbab9cb7d4b8f)