Allow empty instrumentation name
When name is empty, copy the category string into name.

This makes it easier for user to instrument pieces of code that is hard to categorize, in which case, the category and name will mostly be the same string explaining its usage

For example:
If we want to mark the begin and end of VTS test runner's main method to measure the time it used, it does not belong to any of the common categories as it contains framework setup, test setup, and test run.
If empty name is not allowed, the usage will be:
  tfi.Begin('test runner's main method', 'test runner's main method')
which is redundant. Now:
  tfi.Begin('test runner's main method') will have the same effect

Test: python unit test
Bug: 80195850
Change-Id: I5f7a785eea141656703e2d9a1fa258d8f3e30d5e