Modify driver_manager to process arguments or return values that have type TYPE_FMQ_SYNC or TYPE_FMQ_UNSYNC
Test: new feature; add test later
Changes:
* When a function call involves FMQ, driver_manager calls GetQueueDescAddress()
in resource_manager, and receives the address of the queue descriptor.
It stores the address in FunctionSpecificationMessage, which can be
referenced by HAL drivers.
* When a function call returns FMQ, the queue information is stored in
VariableSpecificationMessage. driver_manager passes the message to
resource_manager, who extracts the information and asks fmq_driver to
register the queue object.
* Change the constructor of driver_manager to include resource_manager
pointer as an argument.
* Modify build file to include resource_manager in driver_manager
* Modify SocketServer because the constructor of driver_manager
changes.
* Modify HidlReplayer because the constructor of driver_manager changes.

Change-Id: I1ce2d45db5a6c47c5a3be38fdbe3bcd2ff65971d