Added API in resource_manager and fmq_driver to retrieve queue address
Test: new feature; add test later
Changes:
* resource_manager: new API GetQueueDescAddress(), which allows driver_manager to
preprocess function call arguments with type TYPE_FMQ_SYNC or TYPE_FMQ_UNSYNC.
It returns an address to the queue descriptor, which can be passed in hidl and
be used to create new queue objects.
* resource_manager: new API RegisterFmq(), which takes in a
VariableSpecificationMessage, which contains queue information including
type of data, queue flavor, and queue descriptor address. This method
extracts these information and passes them to fmq_driver.
* fmq_driver: new version of CreateFmq() that can take in a queue descriptor,
and create a new queue object based on the descriptor. This method is
called by resource_manager after it extracts queue information.
* fmq_driver: delete the raw descriptor pointer, because vtsc allocates
it as a raw pointer. Once we create the queue object, we don't need the
descriptor anymore.

Change-Id: I423ec2e5f04e9deebcbe109903df0da06ca02d21