Handle FMQ argument and return type when generating code in vts compiler
Test: new feature; add tests later
Changes:
* When a function call has a fmq argument, check if existing
fmq_desc_address is provided. If provided, cast the fmq_address
back to MessageQueue Descriptor object. If not, create a new queue and
write elements into it.
* When a function call returns a fmq, cast the descriptor pointer
to size_t, and store it in VariableSpecificationMessage, which can be
read back later by driver_manager and passed into resource_manager.
* Update golden/ files.

Change-Id: I547822cfc23eff5dadb64e6bcbfdfd7fada2ec96