Added host-side support of FMQ user-defined type
Test: new feature; add tests later
Changes:
* When a FMQ is returned, determine if it is a scalar type, or
predefined type such as struct, union, or enum. Read scalar_type field
if it is scalar type, and read predefined_type field if it is
user-defined type.
* Allow the option that if a FMQ has scalar type, host side interface
can prepare the data for caller. If it is a user-defined type, caller
has to prepare the write data.

Merged-In: I7bf60ca0d99fb1ff787b05aa69966a7b563374c7
Change-Id: I7bf60ca0d99fb1ff787b05aa69966a7b563374c7