Add shell-default-nohup option for device shell command option
If shell-default-nohup is set to true, all device shell commands will be executed using nohup unless method param nohup (to be added) is set to False.

This CL only adds the shell-default-nohup and related class variables. The implementation of this option and method param nohup will be in other CLs.

Test: run vts
Change-Id: I034acbdfac8eb4d27da3c254c2f9c47b21185c28