To support ITestSuite, do not report SKIP tests to TF.
Previous TF version will ignore test result if testEnded is not called. New TF will treat such case as Failure. Since neither support SKIP, and testAssumptionFailure and testIgnore will both be reported as PASS, the only way to deal with VTS SKIP results is not to report them to TF.

Test: presubmit with and without ITestSuite
Change-Id: Ia89cc402db50b0cf2e0bc4a8d4d11985a71be165